+++ app/src/model/tasktouch.go
@@ -9,9 +9,9 @@ var touchTypes = []string{"SAVED", "DELETED", "COMPLETED", "DISMISSED", "START_U
 
 //TaskTouch is an instance of a user updating or interacting with a Task.
 type TaskTouch struct {
-	ID                int32   `json:"id"`
-	UserID            int32   `json:"userId"`
-	TaskID            int32   `json:"taskId"`
+	ID                int64   `json:"id"`
+	UserID            int64   `json:"userId"`
+	TaskID            int64   `json:"taskId"`
 	TouchTimeStamp    int64   `json:"touchTimeStamp"`
 	LocationTimeStamp int64   `json:"locationTimeStamp"`
 	Longitude         float64 `json:"longitude"`
